feat: add error handling and validation use - update azure,aws,gcp,redis,sql adapters

+++ packages/core/src/adapters/storage.ts
@@ -1,3 +1,4 @@
+// oxlint-disable max-classes-per-file
 // oxlint-disable no-await-in-loop
 // oxlint-disable max-params
 // oxlint-disable require-await
@@ -8,6 +9,8 @@ import * as fsp from "node:fs/promises";
 import * as path from "node:path";
 import { Readable, type Stream } from "node:stream";
 import type { ReadableStream as WebReadableStream } from "node:stream/web";
+import { HTTPException } from "hono/http-exception";
+import type { ContentfulStatusCode } from "hono/utils/http-status";
 import type { LoggerAdapter } from "./logger";
 
 /**
@@ -19,13 +22,18 @@ import type { LoggerAdapter } from "./logger";
  *
  * - `container`: A container/group/bucket to hold files. Each project has one container.
  * - `file`: A single binary that can individually stored and retrieved.
+ *
+ * @throws {StorageNotInitializedError} if the Storage service is not connected.
+ * @throws {ContainerAlreadyExistsError} if the container already exists.
+ * @throws {ContainerDoesNotExistError} if the container does not exist.
+ * @throws {FileDoesNotExistError} if the file does not exist in the container.
  */
 export interface StorageAdapter {
   /**
    * An optional method that is called on app boot-up
    * to run async setup functions.
    * @param options Common options like abortSignal.
-   * @throws if an error occur during initialisation.
+   * @throws If the Storage service fails to initialize.
    */
   init?: (options: StorageAdapterOptions) => Promise<void>;
 
@@ -35,7 +43,7 @@ export interface StorageAdapter {
    * List all containers available in the storage.
    * @param options Common options like abortSignal.
    * @returns A list of names/IDs of the containers.
-   * @throws If the Storage service is not connected.
+   * @throws {StorageNotInitializedError} if the Storage service is not connected.
    */
   listContainers: (options: StorageAdapterOptions) => Promise<string[]>;
 
@@ -148,6 +156,58 @@ export interface StoryBookerFile {
 }
 
 /**
+ * Pre-defined Storage adapter errors
+ * that can be used across different adapters.
+ *
+ * Throws {HTTPException} with relevant status codes.
+ */
+export const StorageAdapterErrors = {
+  StorageNotInitializedError: class extends HTTPException {
+    constructor(cause?: unknown) {
+      super(500, { cause, message: "Storage adapter is not initialized." });
+    }
+  },
+  ContainerAlreadyExistsError: class extends HTTPException {
+    constructor(containerId: string, cause?: unknown) {
+      super(409, {
+        cause,
+        message: `Storage container '${containerId}' already exists.`,
+      });
+    }
+  },
+  ContainerDoesNotExistError: class extends HTTPException {
+    constructor(containerId: string, cause?: unknown) {
+      super(404, {
+        cause,
+        message: `Storage container '${containerId}' does not exist.`,
+      });
+    }
+  },
+  FileDoesNotExistError: class extends HTTPException {
+    constructor(containerId: string, filepath: string, cause?: unknown) {
+      super(404, {
+        cause,
+        message: `Storage file '${filepath}' does not exist in container '${containerId}'.`,
+      });
+    }
+  },
+  FileMalformedError: class extends HTTPException {
+    constructor(containerId: string, filepath: string, cause?: unknown) {
+      super(415, {
+        cause,
+        message: `Storage file '${filepath}' is malformed in container '${containerId}'.`,
+      });
+    }
+  },
+  CustomError: class extends HTTPException {
+    constructor(status: number | undefined, message: string, cause?: unknown) {
+      super(status as ContentfulStatusCode, { cause, message });
+    }
+  },
+  // oxlint-disable-next-line no-explicit-any
+} satisfies Record<string, new (...args: any[]) => HTTPException>;
+
+/**
  * Storage adapter for StoryBooker while uses
  * the local filesystem to read from and write files to.
  * It uses NodeJS FS API to read/write to filesystem.
@@ -178,12 +238,23 @@ export class LocalFileStorage implements StorageAdapter {
 
   createContainer: StorageAdapter["createContainer"] = async (
     containerId,
-    _options,
+    options,
   ) => {
+    if (await this.hasContainer(containerId, options)) {
+      throw new StorageAdapterErrors.ContainerAlreadyExistsError(containerId);
+    }
+
     await fsp.mkdir(this.#genPath(containerId), { recursive: true });
   };
 
-  deleteContainer: StorageAdapter["deleteContainer"] = async (containerId) => {
+  deleteContainer: StorageAdapter["deleteContainer"] = async (
+    containerId,
+    options,
+  ) => {
+    if (!(await this.hasContainer(containerId, options))) {
+      throw new StorageAdapterErrors.ContainerDoesNotExistError(containerId);
+    }
+
     await fsp.rm(this.#genPath(containerId), { force: true, recursive: true });
   };
 
@@ -192,8 +263,15 @@ export class LocalFileStorage implements StorageAdapter {
   };
 
   listContainers: StorageAdapter["listContainers"] = async () => {
+    const dirPath = this.#genPath();
+    if (!fs.existsSync(dirPath)) {
+      throw new StorageAdapterErrors.StorageNotInitializedError(
+        `Dir "${dirPath}" does not exist`,
+      );
+    }
+
     const containers: string[] = [];
-    const entries = await fsp.readdir(this.#genPath(), {
+    const entries = await fsp.readdir(dirPath, {
       withFileTypes: true,
     });
     for (const entry of entries) {
@@ -231,7 +309,15 @@ export class LocalFileStorage implements StorageAdapter {
   downloadFile: StorageAdapter["downloadFile"] = async (
     containerId,
     filepath,
+    options,
   ) => {
+    if (!(await this.hasFile(containerId, filepath, options))) {
+      throw new StorageAdapterErrors.FileDoesNotExistError(
+        containerId,
+        filepath,
+      );
+    }
+
     const path = this.#genPath(containerId, filepath);
     const buffer = await fsp.readFile(path);
     const content = new Blob([buffer as Buffer<ArrayBuffer>]);
